+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 > Features:
+- [64f29b8] [Feature] MCP tools support with OpenAI Assistant mode
 - [440fc19] [Feature] Add gemini-2.0-flash-thinking-exp
 - [e258b78] [Feature] Anthropic Prompt caching fully supported
 
@@ -8,7 +9,6 @@
 > Bug Fixes:
 - [4488ec1] [Bugfix] Trigger mcp servers shutdown before exit
 - [6d4d94e] [Bugfix] Skip asking to save chat on exit if disabled in config
-- [ca1d8c0] [BUGFIX] Re-Fix a bug introduced in 0c7f1af
 
 
 ---